+++ CHANGELOG.md
@@ -12,9 +12,13 @@ This changelog consists of the bug & security fixes and new features being inclu
 
 * #10834 [enhancement] - Refined the admin pages and made approximately 95% of them responsive.
 
-* #10836 [fixed] - Fixed the date of birth issue.
+* #10897 [fixed] - Fixed missing success message when admin approving the review.
+
+* #10883 [fixed] - Fixed broken artisan commands. 
 
 * #10838 [fixed] - Fixed the cart issue.
+
+* #10836 [fixed] - Fixed the date of birth issue.
 
 * #10711 [fixed] - Show validation error in cart rule coupons when entering wrong input.